pyannote_server/requirements.txt
fastapi
uvicorn
pyannote.audio
torch
python-multipart
